optimize transform component world position

+++ src/Engine/Utils/Random.h
@@ -9,7 +9,7 @@ static double randomDouble(double min, double max) {
     static std::random_device rd;
     static std::mt19937 gen(rd());
 
-    std::uniform_real_distribution<float> dis(min, max);
+    std::uniform_real_distribution<double> dis(min, max);
     return dis(gen);
 }
